refactor attribute types
