+++ package.json
@@ -53,6 +53,7 @@
     "rollup-plugin-typescript2": "^0.30.0",
     "ts-jest": "^26.5.6",
     "ts-node": "^9.1.1",
+    "tslib": "^2.3.1",
     "typedoc": "^0.22.4",
     "typescript": "^4.1.3"
   }
