+++ src/domain/knowledge/attribute/index.ts
@@ -1,0 +1,2 @@
+export * from './assets';
+export * from './types';
