--- src/domain/knowledge/attributes/index.ts
@@ -1,3 +1,0 @@
-export * from './key';
-export * from './type';
-export * from './mapping';
